Updating names and fixing boost incompatibility (#3)
* officially seperated from litecoin chain

* block spacing and circulation

* slight changes to block settings and merkle root

* Make boost::multi_index comparators const

Fixes compatibility with boost 1.66

* Replacing litecoin names

Replacing remaining instances of litecoin

* missed one permutation of litecoin, all-caps case. fixed

* Removing broken gen hash code

Code references things that don't exist and are unnecessary.

* Removing broken merkle root code

Broken code references "block" which doesn't exist.

* Replacing code mistakenly removed

* Renaming files

litecoin -> garlicoin

* Changing ports

Changing connection and RPC ports from litecoin defaults

* Testing Chain Genesis Block

Generated genesis block for the testing chain of garlicoin. (Note the testnet has not been set up and this is not that actual garlicoin blockchain)

* Added a seed node

This is the address of the seed node for the test chain. This should be the last piece necessary for the test chain to go live.

* Correcting unix build instructions