Updating POW, Difficulty Adjust and more (#9)
* officially seperated from litecoin chain

* block spacing and circulation

* slight changes to block settings and merkle root

* Make boost::multi_index comparators const

Fixes compatibility with boost 1.66

* Replacing litecoin names

Replacing remaining instances of litecoin

* missed one permutation of litecoin, all-caps case. fixed

* Removing broken gen hash code

Code references things that don't exist and are unnecessary.

* Removing broken merkle root code

Broken code references "block" which doesn't exist.

* Replacing code mistakenly removed

* Renaming files

litecoin -> garlicoin

* Changing ports

Changing connection and RPC ports from litecoin defaults

* Testing Chain Genesis Block

Generated genesis block for the testing chain of garlicoin. (Note the testnet has not been set up and this is not that actual garlicoin blockchain)

* Added a seed node

This is the address of the seed node for the test chain. This should be the last piece necessary for the test chain to go live.

* Correcting unix build instructions

* Moving genesis block to testnet

* Fixing testnet ports

Testnet ports didn't work, a they were generated through find and replace and were too large.

* Correcting missed testnet ports

Missed a few ports in the previous update.

* Added testnet chainTxData

* Fixed regtest merkleRoot

* Make IsInitialBlockDownload return false

This should be removed asap. It is necessary for getting the network up and running.

* Changing min work in test parameters

* This can be removed now

* Proof of Work Change

* Fixes and changes to new PoW algo

* TESTNET HARD FORK SCHEDULED

* Fixing Nfactor implementation

Nfactor incorrectly implemented leading to problems validating blocks (note: correctly implemented for mainnet - this commit must be reverted/modified upon launch)

* Rename scrypt methods to a general form, fix SSE2

Tests using SSE2 are also now functional, although they only test scrypt-n with an Nfactor of 9.

* Remove temporary debug messages

* Added DNS Seed

* Dark Gravity Wave

Change to difficulty adjustment algorithm. Testnet fork scheduled for block 30000

* Bump version before testnet hardfork

* Change version number

* Update chainparams.cpp

* Update garlicoin-tx.1